requirements/base.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.7.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
dj-database-url==0.5.0
Django==2.2
django-celery-beat==2.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-ratelimit==2.0.0
django-timezone-field==5.0
geoip2==2.9.0
gunicorn==19.9.0
html2text==2018.1.9
idna==2.8
kombu==5.2.4
maxminddb==2.2.0
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.1
pytz==2023.3
raven==6.10.0
redis==4.5.4
requests==2.21.0
six==1.16.0
importlib-metadata==4.13.0
python-dotenv
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==4.1.2
